% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{detect_image}
\alias{detect_image}
\title{Detect Objects in Image}
\usage{
detect_image(model, image_file)
}
\arguments{
\item{model}{A loaded model object.}

\item{image_file}{Path to the image file.}
}
\value{
A list containing detection results.
}
\description{
Runs detection on a single image.
}
